backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyJWT==2.8.0
vietnam-provinces
unidecode
google-auth
google-auth-oauthlib